deleteScheduleAction Exception

+++ action/deleteScheduleAction.jsp
@@ -6,19 +6,32 @@
 <%@ page import="java.sql.ResultSet" %>
 
 <% 
-    String scheduleIdx = request.getParameter("scheduleIdx");
-    String year = request.getParameter("year");
-    String month = request.getParameter("month");
-    String day = request.getParameter("day");
+    String year = null;
+    String month = null;
+    String day = null;
+    String errorMessage = null;
 
-    Class.forName("org.mariadb.jdbc.Driver");
-    Connection connect = DriverManager.getConnection("jdbc:mariadb://localhost:3306/calender", "stageus", "1234");
+    try {
+        String scheduleIdx = request.getParameter("scheduleIdx");
 
-    String sql = "DELETE FROM schedule WHERE idx=?;";
-    PreparedStatement query = connect.prepareStatement(sql);
-    query.setString(1, scheduleIdx);
+        if (scheduleIdx == null) {
+            throw new Exception("새로고침 후 다시 이용해주세요");
+        }
+        year = request.getParameter("year");
+        month = request.getParameter("month");
+        day = request.getParameter("day");
 
-    query.executeUpdate();
+        Class.forName("org.mariadb.jdbc.Driver");
+        Connection connect = DriverManager.getConnection("jdbc:mariadb://localhost:3306/calender", "stageus", "1234");
+
+        String sql = "DELETE FROM schedule WHERE idx=?;";
+        PreparedStatement query = connect.prepareStatement(sql);
+        query.setString(1, scheduleIdx);
+
+        query.executeUpdate();
+    } catch (Exception e) {
+        errorMessage = e.getMessage();
+    }
 %>
 
 <head>
@@ -27,8 +40,14 @@
     <title>Document</title>
 </head>
 <body>
-    <script>
-        alert("게시글이 삭제되었습니다")
-        location.href = "../page/detailCalenderPage.jsp?year=<%=year%>&month=<%=month%>&day=<%=day%>"
-    </script>
+    <% if (errorMessage == null) { %>
+        <script>
+            alert("게시글이 삭제되었습니다")
+            location.href = "../page/detailCalenderPage.jsp?year=<%=year%>&month=<%=month%>&day=<%=day%>"
+        </script>
+    <% } else { %>
+        <script>
+            alert("<%=errorMessage%>")
+        </script>
+    <% } %>
 </body>
